be/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
boto3==1.35.19
google-generativeai==0.8.1
apscheduler==3.10.4
python-multipart==0.0.9
